Add the initial implementation of XmlReader
The attributes API is still missing, and a lot of optimizations and
preloading are possible for the elements API.

The error state is not checked in every place, and the actual code
should be more robust with regards to nodes.

All of this is just a matter of adding the unit tests and start
throwing bad stuff at it.