Require an explicit read_end_element to exit the error state
Instead of silently ignoring an error state, use it consistently in
every operation until the developer explicitly exits the current,
invalid "node", by calling xml_reader_read_end_element().

The read_end_element() method unsets any previsouly set error state
and reverts the cursor to the current parent, or the root node if
needed.

This increases the reliability of the API and the error recovery.